AI: some methods. AIRunner: new constructor that takes AI instance. Event: getTime() now returns long. Game: mainLoop implementation. GameQueue: fixed to fit new Event contract. GraphicsEngine: added init() method. Player: now manages a AIRunner. Skill: fixed to fit new Event contract. Updater: many implementations.

+++ trunk/PA/src/Skill.java
@@ -120,7 +120,7 @@ public enum Skill {
 	},
 	Froze(false) {
 		public void use(Game game, Player src, Player target) {
-			int time_effect_start = game.getTime() + 3000;
+			long time_effect_start = game.getTime() + 3000;
 			
 			// Schedule events to apply frozen effect to all players except itself
 			GameQueue queue = game.getGameQueue();
